Update the webpage content and JavaScript code to support the final version of the app and load Gemini Nano.
* **README.md**
  - Update the custom domain reference to `Ryukijano.github.io`.
  - Include the project description and challenge description.
  - Remove sections related to Chrome AI API documentation, setup instructions, GitHub Pages setup, Early Preview Program, feedback submission, and demonstration video.

* **app.js**
  - Update the JavaScript code to support the final version of the app.
  - Include functionality to load Gemini Nano when the website is opened.

* **index.html**
  - Add a section to load Gemini Nano when opening the website.
  - Update the project repository link.

* **manifest.json**
  - Remove references to the Chrome Extension.
  - Ensure it reflects the final version of the app as a webpage.

* **background.js** and **content.js**
  - Delete these files.
