+++ app.js
@@ -1,211 +1,32 @@
-/**
- * Copyright 2024 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { marked } from "https://cdn.jsdelivr.net/npm/marked@13.0.3/lib/marked.esm.js";
-import DOMPurify from "https://cdn.jsdelivr.net/npm/dompurify@3.1.6/dist/purify.es.mjs";
-
-(async () => {
-  const errorMessage = document.getElementById("error-message");
-  const costSpan = document.getElementById("cost");
-  const promptArea = document.getElementById("prompt-area");
-  const problematicArea = document.getElementById("problematic-area");
-  const promptInput = document.getElementById("prompt-input");
-  const responseArea = document.getElementById("response-area");
-  const copyLinkButton = document.getElementById("copy-link-button");
-  const resetButton = document.getElementById("reset-button");
-  const copyHelper = document.querySelector("small");
-  const rawResponse = document.querySelector("details div");
-  const form = document.querySelector("form");
-  const maxTokensInfo = document.getElementById("max-tokens");
-  const temperatureInfo = document.getElementById("temperature");
-  const tokensLeftInfo = document.getElementById("tokens-left");
-  const tokensSoFarInfo = document.getElementById("tokens-so-far");
-  const topKInfo = document.getElementById("top-k");
-  const sessionTemperature = document.getElementById("session-temperature");
-  const sessionTopK = document.getElementById("session-top-k");
-
-  responseArea.style.display = "none";
-
-  let session = null;
-
-  if (!self.ai || !self.ai.languageModel) {
-    errorMessage.style.display = "block";
-    errorMessage.innerHTML = `Your browser doesn't support the Prompt API. If you're on Chrome, join the <a href="https://developer.chrome.com/docs/ai/built-in#get_an_early_preview">Early Preview Program</a> to enable it.`;
-    return;
-  }
-
-  promptArea.style.display = "block";
-  copyLinkButton.style.display = "none";
-  copyHelper.style.display = "none";
-
-  const promptModel = async (highlight = false) => {
-    copyLinkButton.style.display = "none";
-    copyHelper.style.display = "none";
-    problematicArea.style.display = "none";
-    const prompt = promptInput.value.trim();
-    if (!prompt) return;
-    responseArea.style.display = "block";
-    const heading = document.createElement("h3");
-    heading.classList.add("prompt", "speech-bubble");
-    heading.textContent = prompt;
-    responseArea.append(heading);
-    const p = document.createElement("p");
-    p.classList.add("response", "speech-bubble");
-    p.textContent = "Generating response...";
-    responseArea.append(p);
-    let fullResponse = "";
-
-    try {
-      if (!session) {
-        await updateSession();
-        updateStats();
-      }
-      const stream = await session.promptStreaming(prompt);
-
-      for await (const chunk of stream) {
-        fullResponse = chunk.trim();
-        p.innerHTML = DOMPurify.sanitize(marked.parse(fullResponse));
-        rawResponse.innerText = fullResponse;
-      }
-    } catch (error) {
-      p.textContent = `Error: ${error.message}`;
-    } finally {
-      if (highlight) {
-        problematicArea.style.display = "block";
-        problematicArea.querySelector("#problem").innerText =
-          decodeURIComponent(highlight).trim();
-      }
-      copyLinkButton.style.display = "inline-block";
-      copyHelper.style.display = "inline";
-      updateStats();
+document.addEventListener('DOMContentLoaded', () => {
+    // Function to interact with the Prompt API and send results to the background script
+    async function interactWithPromptAPI(prompt) {
+        try {
+            const { available } = await ai.languageModel.capabilities();
+            if (available !== "no") {
+                const session = await ai.languageModel.create();
+                const result = await session.prompt(prompt);
+                displayResult(result);
+            } else {
+                displayError("AI model is not available.");
+            }
+        } catch (error) {
+            displayError(error.message);
+        }
     }
-  };
 
-  const updateStats = () => {
-    if (!session) {
-      return;
+    // Function to display the result on the webpage
+    function displayResult(result) {
+        const featureOutput = document.getElementById('feature-output');
+        featureOutput.innerHTML = `<p>${result}</p>`;
     }
-    const { maxTokens, temperature, tokensLeft, tokensSoFar, topK } = session;
-    maxTokensInfo.textContent = new Intl.NumberFormat("en-US").format(
-      maxTokens,
-    );
-    (temperatureInfo.textContent = new Intl.NumberFormat("en-US", {
-      maximumSignificantDigits: 5,
-    }).format(temperature)),
-      (tokensLeftInfo.textContent = new Intl.NumberFormat("en-US").format(
-        tokensLeft,
-      ));
-    tokensSoFarInfo.textContent = new Intl.NumberFormat("en-US").format(
-      tokensSoFar,
-    );
-    topKInfo.textContent = new Intl.NumberFormat("en-US").format(topK);
-  };
 
-  const params = new URLSearchParams(location.search);
-  const urlPrompt = params.get("prompt");
-  const highlight = params.get("highlight");
-  if (urlPrompt) {
-    promptInput.value = decodeURIComponent(urlPrompt).trim();
-    await promptModel(highlight);
-  }
-
-  form.addEventListener("submit", async (e) => {
-    e.preventDefault();
-    await promptModel();
-  });
-
-  promptInput.addEventListener("keydown", (e) => {
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-      form.dispatchEvent(new Event("submit"));
+    // Function to display error messages on the webpage
+    function displayError(message) {
+        const featureOutput = document.getElementById('feature-output');
+        featureOutput.innerHTML = `<p class="error">${message}</p>`;
     }
-  });
 
-  promptInput.addEventListener("focus", () => {
-    promptInput.select();
-  });
-
-  promptInput.addEventListener("input", async () => {
-    const value = promptInput.value.trim();
-    if (!value) {
-      return;
-    }
-    const cost = await session.countPromptTokens(value);
-    if (!cost) {
-      return;
-    }
-    costSpan.textContent = `${cost} token${cost === 1 ? '' : 's'}`;
-  });
-
-  const resetUI = () => {
-    responseArea.style.display = "none";
-    responseArea.innerHTML = "";
-    rawResponse.innerHTML = "";
-    problematicArea.style.display = "none";
-    copyLinkButton.style.display = "none";
-    copyHelper.style.display = "none";
-    maxTokensInfo.textContent = "";
-    temperatureInfo.textContent = "";
-    tokensLeftInfo.textContent = "";
-    tokensSoFarInfo.textContent = "";
-    topKInfo.textContent = "";
-    promptInput.focus();
-  };
-
-  resetButton.addEventListener("click", () => {
-    promptInput.value = "";
-    resetUI();
-    session.destroy();
-    session = null;
-    updateSession();
-  });
-
-  copyLinkButton.addEventListener("click", () => {
-    const prompt = promptInput.value.trim();
-    if (!prompt) return;
-    const url = new URL(self.location.href);
-    url.searchParams.set("prompt", encodeURIComponent(prompt));
-    const selection = getSelection().toString() || "";
-    if (selection) {
-      url.searchParams.set("highlight", encodeURIComponent(selection));
-    } else {
-      url.searchParams.delete("highlight");
-    }
-    navigator.clipboard.writeText(url.toString()).catch((err) => {
-      alert("Failed to copy link: ", err);
-    });
-    const text = copyLinkButton.textContent;
-    copyLinkButton.textContent = "Copied";
-    setTimeout(() => {
-      copyLinkButton.textContent = text;
-    }, 3000);
-  });
-
-  const updateSession = async () => {
-    session = await self.ai.languageModel.create({
-      temperature: Number(sessionTemperature.value),
-      topK: Number(sessionTopK.value),
-    });
-    resetUI();
-    updateStats();
-  };
-
-  sessionTemperature.addEventListener("input", async () => {
-    await updateSession();
-  });
-
-  sessionTopK.addEventListener("input", async () => {
-    await updateSession();
-  });
-
-  if (!session) {
-    const { defaultTopK, maxTopK, defaultTemperature } =
-      await self.ai.languageModel.capabilities();
-    sessionTemperature.value = defaultTemperature;
-    sessionTopK.value = defaultTopK;
-    sessionTopK.max = maxTopK;
-    await updateSession();
-  }
-})();
+    // Load Gemini Nano when the website is opened
+    interactWithPromptAPI("Load Gemini Nano");
+});
