--- docs/demo.md
@@ -1,20 +1,0 @@
-# Demonstration Video
-
-## Video Link
-
-Watch the demonstration video to see the project in action: [Demonstration Video](https://www.youtube.com/watch?v=example)
-
-## Instructions for Creating and Uploading the Demonstration Video
-
-1. **Create the Video:**
-   - Record a video (about 3 minutes) that demonstrates your project and its functionality.
-   - Ensure the video clearly shows how your web application or Chrome Extension uses the built-in AI APIs.
-   - Highlight key features and any unique aspects of your project.
-
-2. **Upload the Video:**
-   - Upload the video to YouTube or Vimeo.
-   - Make sure the video is set to public so that it can be accessed by the judges and other participants.
-
-3. **Share the Video Link:**
-   - Provide the URL to your public video in your project submission.
-   - Include the video link in your project's README.md file and any other relevant documentation.
