--- docs/feedback.md
@@ -1,10 +1,0 @@
-# Feedback Submission and Contact Information
-
-We value your feedback! Please share your thoughts and experiences with us to help refine and improve the APIs and models.
-
-- **Feedback form for quality or technical issues:** [Feedback Form](https://goo.gle/chrome-ai-dev-preview-feedback-quality)
-- **Report bugs or issues related to Chrome’s behavior / implementation of the Prompt API:** [Chromium Bug Report](https://bugs.chromium.org/)
-- **Report ergonomic issues or other problems related to one of the built-in AI APIs:**
-  - [Prompt API spec issues](https://github.com/WICG/prompt-api/issues)
-  - [Translation API spec issues](https://github.com/WICG/translation-api/issues)
-- **Other questions or issues:** Reach out directly by sending an email to the mailing list owners (chrome-ai-dev-preview+owners@chromium.org).
