+++ index.html
@@ -16,9 +16,13 @@
             <h2>AI-Powered Features</h2>
             <div id="feature-output"></div>
         </section>
+        <section id="gemini-nano">
+            <h2>Gemini Nano</h2>
+            <p>Loading Gemini Nano...</p>
+        </section>
     </main>
     <footer>
-        <p><a href="https://github.com/your-repository">Project Repository</a></p>
+        <p><a href="https://github.com/google-chrome-ai-gemini-nano">Project Repository</a></p>
         <p><a href="https://goo.gle/chrome-ai-dev-preview-feedback-quality">Submit Feedback</a></p>
     </footer>
     <script src="app.js"></script>
