--- styles.css
@@ -1,39 +1,0 @@
-/* Basic styles for the web application layout */
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f4f4f4;
-}
-
-header {
-    background-color: #4CAF50;
-    color: white;
-    padding: 1em;
-    text-align: center;
-}
-
-main {
-    padding: 1em;
-}
-
-footer {
-    background-color: #333;
-    color: white;
-    text-align: center;
-    padding: 1em;
-    position: fixed;
-    width: 100%;
-    bottom: 0;
-}
-
-/* Responsive design rules for different screen sizes */
-@media (max-width: 600px) {
-    header, footer {
-        padding: 0.5em;
-    }
-
-    main {
-        padding: 0.5em;
-    }
-}
